// src/login/dto/jwt-decodeClient.dto.ts
export class JwtDecodeClient {
  id: string;
  name: string;
  cpf_cnpj: string;
  client_type: string;
  birthday: Date;
  phone: string;
  email: string;
  password: string;
  zipCode: string;
  street: string;
  number: string;
  neighbourhood: string;
  city: string;
  iat: number;
  exp: number;
}